Update to Vaser 1.1.0-alpha2

+++ VaserChat/SharedCodebase/PACKETS.cs
@@ -10,7 +10,7 @@ namespace Global
     public class SEND_LOGIN : Container
     {
         public static SEND_LOGIN slCont = new SEND_LOGIN();
-        internal const int ContID = 1;
+        internal const ushort ContID = 1;
 
 
         public string Username = "NONE";
@@ -19,18 +19,18 @@ namespace Global
     public class SEND_MESSAGE : Container
     {
         public static SEND_MESSAGE smCont = new SEND_MESSAGE();
-        internal const int ContID = 2;
+        internal const ushort ContID = 2;
 
-        public int UserID = 0;
+        public uint UserID = 0;
         public string Message = "NONE";
     }
 
     public class SEND_USER : Container
     {
         public static SEND_USER suCont = new SEND_USER();
-        internal const int ContID = 3;
+        internal const ushort ContID = 3;
 
-        public int UserID = 0;
+        public uint UserID = 0;
         public string Username = "NONE";
         public int StatusID = 0;
     }
